Change tests to be more unit-ish
